+++ attributes/5.6.rb
@@ -1,11 +1,16 @@
-default['codenamephp_php']['5.6']['package_name']['cli'] = 'php5.6-cli'
-default['codenamephp_php']['5.6']['package_name']['apache'] = 'php5.6'
+default['codenamephp_php']['5.6']['package_name'] = 'php5.6-cli'
 default['codenamephp_php']['5.6']['additional_packages'] = [
-  'php5.6-xml',
-  'php5.6-tidy',
+  'php5.6-bcmath',
   'php5.6-curl',
   'php5.6-gd',
+  'php5.6-json',
+  'php5.6-mcrypt',
   'php5.6-mbstring',
+  'php5.6-mysql',
+  'php5.6-tidy',
+  'php5.6-xml',
+  'php5.6-xsl',
   'php5.6-xmlrpc',
-  'php5.6-mysql'
+  'php5.6-zip',
+  'php-imagick'
 ]