+++ attributes/7.2.rb
@@ -1,12 +1,15 @@
-default['codenamephp_php']['7.2']['package_name']['cli'] = 'php7.2-cli'
-default['codenamephp_php']['7.2']['package_name']['apache'] = 'php7.2'
-default['codenamephp_php']['7.2']['additional_packages'] = ['php7.2-xml']
+default['codenamephp_php']['7.2']['package_name'] = 'php7.2-cli'
 default['codenamephp_php']['7.2']['additional_packages'] = [
-  'php7.2-xml',
-  'php7.2-tidy',
+  'php7.2-bcmath',
   'php7.2-curl',
   'php7.2-gd',
+  'php7.2-json',
   'php7.2-mbstring',
+  'php7.2-mysql',
+  'php7.2-tidy',
+  'php7.2-xml',
+  'php7.2-xsl',
   'php7.2-xmlrpc',
-  'php7.2-mysql'
+  'php7.2-zip',
+  'php-imagick'
 ]