+++ test/smoke/5.6/5.6_test.rb
@@ -7,11 +7,7 @@ control 'php-5.6-1.0' do
   title 'Make sure php5.6 is intsalled'
   desc 'Install php5.6 and apache2'
 
-  describe package('php5.6') do
-    it { should be_installed }
-  end
-
-  describe package('apache2') do
+  describe package('php5.6-cli') do
     it { should be_installed }
   end
 end